requirements.txt
fastapi
uvicorn[standard]
numpy
pydantic
requests
python-dotenv
